Introduce `system.rs`
It gains three types from `plumbing.rs` that define the high level query
system structures.
